+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -19,6 +19,10 @@ Notes](../../RELEASENOTES.md).
 * Fix incorrect handling of leading digits in metric names for OpenMetrics.
   ([#7454](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7454))
 
+* Add `PrometheusAspNetCoreOptions.ScopeInfoEnabled` property to enable or
+  disable scope labels in Prometheus metrics. Defaults to `true`.
+  ([#7436](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7436))
+
 * Added support for the `dots` and `values` Prometheus UTF-8 name escaping
   schemes when negotiated via the `Accept` header.
   ([#7439](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7439))